refactor(tests): move selfdestruct bal tests to oog file; add gas boundaries
